#23 WIP(image details): make the image details page for intercepting route
