+++ components/ImageCard.tsx
@@ -41,7 +41,7 @@ export default function ImageCard(props: ImageCardProps) {
   return (
     <>
       <div key={image.url} className="">
-        <Link href={`photos/${image.public_id}`} key={image.public_id}>
+        <Link href={`photos/${image.public_id}`} key={image.public_id} scroll={false}>
           <CCldImage
             src={image.public_id}
             layout="responsive"
